+++ issuetracker/CHANGES.rst
@@ -20,6 +20,7 @@ Other changes
 
   - Fixed TypeError caused by ``launchpad`` issue tracker
   - Fixed issue title in ``launchpad`` issue tracker
+  - Fixed detection of closed issues in ``launchpad`` issue tracker
   - Fixed CSS classes for issue references to be more compatible with Sphinx
     themes